created abstraction class DeletableEntity, made Cart the sublcass of DeletableEntity
